+++ src/CHANGELOG.md
@@ -1,3 +1,11 @@
+- 1.0.6
+  - added `pristine` and `submitting` properties in scope of form
+  - `scrollIntoViewFirstInvalidInputGroup()` gets called only on submit, not with every checkValidity
+- 1.0.5
+  - fixing build
 - 1.0.4
   - Now it queries for Input node again when the node was destroyed by component update
   - Added actions and reducer, so can be connected to Redux store
+
+
+